requirements.txt
requests
beautifulsoup4
telegram
asyncio
python-telegram-bot
botcity-maestro-sdk